+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.0] - 2025-08-20
+
+### Breaking Changes
+
+- Changed the default remote entity name and the entity id for the light source timer sensor
+  - Please re-run the setup when running as an external integration (e.g. via Docker)
+
+### Fixed
+
+- Fixed a timeout while sending an on/off/toggle command ([#2](https://github.com/kennymc-c/ucr-integration-sonyADCP/issues/2))
+- Fixed handling of non supported query commands used for the video signal sensor/media playback attributes
+- Fixed wrong English sensor names
+
+### Changed
+
+- Improved handling of updating attributes for entities that have not been added as configured entities
+
 ## [1.2.0] - 2025-07-21
 
 ### Fixed